xsession. Do not source ~/.profile from ~/.Xsession.
Really, i should never do this. This should be done in system-wide Xsession
files: /etc/gdm3/Xsession does this (and first sources /etc/profile, which is
the right thing) and /etc/X11/Xsession does not, because it is intended to run
by `startx`, which already runs from shell (which source profile files).