fix: build. Fix installation for recent stack versions.
Recent stack versions (supporting extensible snapshots) changed
something.. and now to specify flags for `xmobar` i need to add
`extra-dep` to its specification in `packages`. Though, making `xmobar`
to be a dependency won't work: `xmobar` provides only executable and my
config does not really depends on it (well, it can't, because there is
no library). So, in that case, `xmobar` just won't be built or
installed.

In fact, i want a dependency on executable, which isn't simple to
specify (i didn't find a way).

Though, leaving `xmobar` in `packages` section and specifying
`extra-dep: false` will do the trick: `xmobar` will be built and
installed as i want. Though, this seems not very correct according to
`packages` intended meaning.